Download files posted to Slack using file URL
When using something like IRC with Slack, files that others post to
Slack get rendered as a URL in the IRC chat. I added a `download`
function, as well as a `--download / -d` option to `bin/slackcat` to be
able to download the actual file posted using just the URL.

The file is downloaded to the current directory from where `slackcat` is
run, and is saved with the file name used by Slack.

In the future, I may add an additional option for the user to specify
where to save the downloaded file to.